Feat: finials the linear delete function
This function is used to delete a record from the table. It takes a
hashmap of key-value pairs, where the key is the column name and the
value is the value to be deleted. It then searches for the record
with the given key-value pairs and deletes it. If the record is not
found, it returns an error message. Conjunctivity is assumed between
the key-value pairs. In other words, all key-value pairs must be
satisfied for a record to be deleted.

The next step in this branch, is to implement the update function.