Feat: finished the update method in `DBApp.java`
The current method is using a naive approach to update the table. It
firstly deletes the record and then inserts the new record. The
deletion is done by calling the delete method in `DBApp.java`. The
insertion is done by calling the insert method in `DBApp.java`. This
method is bound to change in the future once the B+ tree is finalized.

This should be the last feature heavy commit in this branch. A merge to
the main branch is expected to happen soon. This does not mean that the
select, delete, and update methods are finalized. They are still subject
to change.