Feat: finalise the delete functionality with testing
The delete functionality has been finalised and tested. Now it will run
successfully even when a tree is not found. Furthermore the test cases
have been updated to reflect this change. Though the test cases are
omitted from this commit, as they are not relevant to the functionality
of the program. Though if needed the developer can add the following
snippet of code to DBAPP.java in the main method at line 64:

```
// testing the delete
htblColNameValue.put("id", Integer.valueOf(9));
dbApp.deleteFromTable(strTableName, htblColNameValue);
```